Significant changes to model and loss function
The previous model utilized a weighted MSE. This is unconventional for most RNNs it results in the optimzer searching through a non-convex space. Therefore, it is vulnerable to issues with local optima. Unexpectedly, a weighted MSE worked very well. I decided to switch to a standard loss and activation function for multi-label classifiation: sigmoid cross entropy and sigmoid activation respectively. Theoretically, this works better as it doesn't have issues related to non-convexity. It turns out in this case, it converges much slower and is much harder to tune. However, this is the standard approach that should work best for most problems. I changed the default hyperparameters to adjust for the change by greatly reducing regularization weights, and increasing the number of epochs. I recommend implementing a beter hyperparameter search algorithm or use SageMaker's HPO feature as I found it harder to achieve great results (customer segmentation as a pre-processing step would also be ideal if you have a larger data set).